% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serex.R
\name{addDataSheet}
\alias{addDataSheet}
\title{writes a dataframe to the chosen sheet,
format it as a "datasheet"}
\usage{
addDataSheet(df, name = "DataSheet", sheet, startRow = 12, startCol = 2,
  topspace = 3)
}
\arguments{
\item{x:}{dataframe to store}

\item{name:}{title of the datasheet to write}

\item{sheet:}{sheet object to write on}

\item{startRow:}{int, row for the upper left corner of the datasheet}

\item{startCol:}{int, column for the upper left corner of the datasheet}

\item{topspace:}{int, space between the top edge of the sheet and the header of the table}
}
\value{
nothing
}
\description{
writes a dataframe to the chosen sheet,
format it as a "datasheet"
}

