% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/borderStyles.R
\name{borderStyles}
\alias{borderStyles}
\title{retirn a set of 4 boder edge styles
and 4 border cornet styles
+ shadow style}
\usage{
borderStyles(wb)
}
\arguments{
\item{workbook}{object}
}
\description{
retirn a set of 4 boder edge styles
and 4 border cornet styles
+ shadow style
}
\seealso{
Other internal: \code{\link{setStyles}},
  \code{\link{xlsx.addTitle}},
  \code{\link{xlsx.sheetborder}}
}
\keyword{style}

