% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serex.R
\name{createSE}
\alias{createSE}
\title{creates new workbook, registers styles
and adds given sheets
NOTE: maybe should return wb only, then get sheets through getSheets()}
\usage{
createSE(sheetlist = list(SUMMARY = "Neighborhood inventory"))
}
\arguments{
\item{sheetList:}{named list, list(SheetName='Sheet title')}
}
\value{
named list of WB (workbook) and Sheets (named list of sheets)

sheets: named list of registered sheets
}
\description{
creates new workbook, registers styles
and adds given sheets
NOTE: maybe should return wb only, then get sheets through getSheets()
}

