% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serex.R
\name{saveSE}
\alias{saveSE}
\title{store SE workbook as an excel file}
\usage{
saveSE(wb, path)
}
\arguments{
\item{wb:}{workbook object to save}

\item{path:}{string, path to the file}
}
\description{
store SE workbook as an excel file
}

