% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStyles.R
\name{setStyles}
\alias{setStyles}
\title{retirn a set of 4 boder edge styles
and 4 border cornet styles
+ shadow style}
\usage{
setStyles(wb)
}
\arguments{
\item{workbook}{object}
}
\value{
named list of styles
}
\description{
retirn a set of 4 boder edge styles
and 4 border cornet styles
+ shadow style
}
\seealso{
Other internal: \code{\link{borderStyles}},
  \code{\link{xl.addTitle}}, \code{\link{xl.sheetborder}}
}
\keyword{style}

