% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setTableRows.r
\name{setTableRows}
\alias{setTableRows}
\title{sets table rows to colored formats,}
\usage{
setTableRows(startRow, startCol, shape, sheet)
}
\arguments{
\item{startRow:}{int, row for the upper left corner of the datasheet}

\item{startCol:}{int, column for the upper left corner of the datasheet}

\item{shape:}{vector of 2, dataframe lengiht and width}

\item{sheet:}{sheet object to write on}
}
\value{
nothing
}
\description{
sets table rows to colored formats,
}

