% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.addTitle.R
\name{xl.addTitle}
\alias{xl.addTitle}
\title{add title to the sheet}
\usage{
xl.addTitle(sheet, rowIndex, colIndex, title, titleStyle)
}
\arguments{
\item{xlsx}{sheet object}

\item{rowIndex:}{vector of integers, difining rows to work on}

\item{colIndex:}{vector of integers, difining columns to work on}

\item{title:}{string, title to write}

\item{titleStyle:}{xlsx style format}
}
\description{
add title to the sheet
}
\seealso{
Other internal: \code{\link{borderStyles}},
  \code{\link{setStyles}}, \code{\link{xl.sheetborder}}
}
\keyword{styler}

