% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.background.R
\name{xl.background}
\alias{xl.background}
\title{apply shadowed background}
\usage{
xl.background(sheet, background_style)
}
\arguments{
\item{xlsx}{sheet object}

\item{background_style:}{xlsx style object}
}
\description{
apply shadowed background
}
\keyword{styler}

