% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl.sheetborder.R
\name{xl.sheetborder}
\alias{xl.sheetborder}
\title{draws a sheet border surrounding the dataframe}
\usage{
xl.sheetborder(sheet, width, height, rowStart, colStart)
}
\arguments{
\item{xlsx}{sheet object}

\item{wight:}{integer, wight of the sheet, edge to edge}

\item{height:}{integer, height of the sheet, edge to edge}

\item{rowStart:}{integer, row for the upper left corner}

\item{colStart:}{integer, col for the upper left corner}
}
\description{
draws a sheet border surrounding the dataframe
}
\seealso{
Other internal: \code{\link{borderStyles}},
  \code{\link{setStyles}}, \code{\link{xl.addTitle}}
}
\keyword{styler}

