% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsx.background.R
\name{xlsx.background}
\alias{xlsx.background}
\title{apply shadowed background}
\usage{
xlsx.background(sheet, background_style)
}
\arguments{
\item{xlsx}{sheet object}

\item{background_style:}{xlsx style object}
}
\description{
apply shadowed background
}
\keyword{styler}

